Struct matrix_crypto::OlmMachine

pub struct OlmMachine { /* private fields */ }

A high level state machine that handles E2EE for Matrix.

Implementations

impl OlmMachine

pub fn new(
user_id: &str,
device_id: &str,
path: &str
) -> Result<Self, CryptoStoreError>

Create a new OlmMachine

Arguments

user_id - The unique ID of the user that owns this machine.

device_id - The unique ID of the device that owns this machine.

path - The path where the state of the machine should be persisted.

pub fn user_id(&self) -> String

Get the user ID of the owner of this OlmMachine.

pub fn device_id(&self) -> String

Get the device ID of the device of this OlmMachine.

pub fn display_name(&self) -> Result<Option<String>, CryptoStoreError>

Get the display name of our own device.

pub fn get_identity(
&self,
user_id: &str
) -> Result<Option<UserIdentity>, CryptoStoreError>

Get a cross signing user identity for the given user ID.

pub fn is_identity_verified(
&self,
user_id: &str
) -> Result<bool, CryptoStoreError>

Check if a user identity is considered to be verified by us.

pub fn verify_identity(
&self,
user_id: &str
) -> Result<SignatureUploadRequest, SignatureError>

Manually the user with the given user ID.

This method will attempt to sign the user identity using either our private cross signing key, for other user identities, or our device keys for our own user identity.

This method can fail if we don’t have the private part of our user-signing key.

Returns a request that needs to be sent out for the user identity to be marked as verified.

pub fn get_device(
&self,
user_id: &str,
device_id: &str
) -> Result<Option<Device>, CryptoStoreError>

Get a Device from the store.

Arguments

user_id - The id of the device owner.

device_id - The id of the device itself.

pub fn verify_device(
&self,
user_id: &str,
device_id: &str
) -> Result<SignatureUploadRequest, SignatureError>

Manually the device of the given user with the given device ID.

This method will attempt to sign the device using our private cross signing key.

This method will always fail if the device belongs to someone else, we can only sign our own devices.

It can also fail if we don’t have the private part of our self-signing key.

Returns a request that needs to be sent out for the device to be marked as verified.

pub fn mark_device_as_trusted(
&self,
user_id: &str,
device_id: &str
) -> Result<(), CryptoStoreError>

Mark the device of the given user with the given device id as trusted.

pub fn get_user_devices(
&self,
user_id: &str
) -> Result<Vec<Device>, CryptoStoreError>

Get all devices of an user.

Arguments

user_id - The id of the device owner.

pub fn identity_keys(&self) -> HashMap<String, String>

Get our own identity keys.

pub fn outgoing_requests(&self) -> Result<Vec<Request>, CryptoStoreError>

Get the list of outgoing requests that need to be sent to the homeserver.

After the request was sent out and a successful response was received the response body should be passed back to the state machine using the mark_request_as_sent() method.

Note: This method call should be locked per call.

pub fn mark_request_as_sent(
&self,
request_id: &str,
request_type: RequestType,
response_body: &str
) -> Result<(), CryptoStoreError>

Mark a request that was sent to the server as sent.

Arguments

request_id - The unique ID of the request that was sent out. This needs to be an UUID.

request_type - The type of the request that was sent out.

response_body - The body of the response that was received.

pub fn receive_sync_changes(
&self,
events: &str,
device_changes: DeviceLists,
key_counts: HashMap<String, i32>,
unused_fallback_keys: Option<Vec<String>>
) -> Result<String, CryptoStoreError>

Let the state machine know about E2EE related sync changes that we received from the server.

This needs to be called after every sync, ideally before processing any other sync changes.

Arguments

events - A serialized array of to-device events we received in the current sync response.

device_changes - The list of devices that have changed in some way since the previous sync.

key_counts - The map of uploaded one-time key types and counts.

pub fn update_tracked_users(&self, users: Vec<String>)

Add the given list of users to be tracked, triggering a key query request for them.

Note: Only users that aren’t already tracked will be considered for an update. It’s safe to call this with already tracked users, it won’t result in excessive keys query requests.

Arguments

users - The users that should be queued up for a key query.

pub fn is_user_tracked(&self, user_id: &str) -> Result<bool, CryptoStoreError>

Check if the given user is considered to be tracked.

A user can be marked for tracking using the OlmMachine::update_tracked_users() method.

pub fn get_missing_sessions(
&self,
users: Vec<String>
) -> Result<Option<Request>, CryptoStoreError>

Generate one-time key claiming requests for all the users we are missing sessions for.

After the request was sent out and a successful response was received the response body should be passed back to the state machine using the mark_request_as_sent() method.

This method should be called every time before a call to share_group_session() is made.

Arguments

users - The list of users for which we would like to establish 1:1 Olm sessions for.

pub fn share_room_key(
&self,
room_id: &str,
users: Vec<String>
) -> Result<Vec<Request>, CryptoStoreError>

Share a room key with the given list of users for the given room.

After the request was sent out and a successful response was received the response body should be passed back to the state machine using the mark_request_as_sent() method.

This method should be called every time before a call to encrypt() with the given room_id is made.

Arguments

room_id - The unique id of the room, note that this doesn’t strictly need to be a Matrix room, it just needs to be an unique identifier for the group that will participate in the conversation.

users - The list of users which are considered to be members of the room and should receive the room key.

pub fn encrypt(
&self,
room_id: &str,
event_type: &str,
content: &str
) -> Result<String, CryptoStoreError>

Encrypt the given event with the given type and content for the given room.

Note: A room key needs to be shared with the group of users that are members in the given room. If this is not done this method will panic.

The usual flow to encrypt an event using this state machine is as follows:

Get the one-time key claim request to establish 1:1 Olm sessions for the room members of the room we wish to participate in. This is done using the get_missing_sessions() method. This method call should be locked per call.

Share a room key with all the room members using the share_group_session(). This method call should be locked per room.

Encrypt the event using this method.

Send the encrypted event to the server.

After the room key is shared steps 1 and 2 will become noops, unless there’s some changes in the room membership or in the list of devices a member has.

Arguments

room_id - The unique id of the room where the event will be sent to.

even_type - The type of the event.

content - The serialized content of the event.

pub fn decrypt_room_event(
&self,
event: &str,
room_id: &str
) -> Result<DecryptedEvent, DecryptionError>

Decrypt the given event that was sent in the given room.

Arguments

event - The serialized encrypted version of the event.

room_id - The unique id of the room where the event was sent to.

pub fn request_room_key(
&self,
event: &str,
room_id: &str
) -> Result<KeyRequestPair, DecryptionError>

Request or re-request a room key that was used to encrypt the given event.

Arguments

event - The undecryptable event that we would wish to request a room key for.

room_id - The id of the room the event was sent to.

pub fn export_keys(
&self,
passphrase: &str,
rounds: i32
) -> Result<String, CryptoStoreError>

Export all of our room keys.

Arguments

passphrase - The passphrase that should be used to encrypt the key export.

rounds - The number of rounds that should be used when expanding the passphrase into an key.

pub fn import_keys(
&self,
keys: &str,
passphrase: &str,
progress_listener: Box<dyn ProgressListener>
) -> Result<KeysImportResult, KeyImportError>

Import room keys from the given serialized key export.

Arguments

keys - The serialized version of the key export.

passphrase - The passphrase that was used to encrypt the key export.

progress_listener - A callback that can be used to introspect the progress of the key import.

pub fn import_decrypted_keys(
&self,
keys: &str,
progress_listener: Box<dyn ProgressListener>
) -> Result<KeysImportResult, KeyImportError>

Import room keys from the given serialized unencrypted key export.

This method is the same as OlmMachine::import_keys but the decryption step is skipped and should be performed by the caller. This should be used if the room keys are coming from the server-side backup, the method will mark all imported room keys as backed up.

Arguments

keys - The serialized version of the unencrypted key export.

progress_listener - A callback that can be used to introspect the progress of the key import.

pub fn discard_room_key(&self, room_id: &str) -> Result<(), CryptoStoreError>

Discard the currently active room key for the given room if there is one.

pub fn get_verification_requests(
&self,
user_id: &str
) -> Vec<VerificationRequest>

Get all the verification requests that we share with the given user.

Arguments

user_id - The ID of the user for which we would like to fetch the verification requests.

pub fn get_verification_request(
&self,
user_id: &str,
flow_id: &str
) -> Option<VerificationRequest>

Get a verification requests that we share with the given user with the given flow id.

Arguments

user_id - The ID of the user for which we would like to fetch the verification requests.

flow_id - The ID that uniquely identifies the verification flow.

pub fn accept_verification_request(
&self,
user_id: &str,
flow_id: &str,
methods: Vec<String>
) -> Option<OutgoingVerificationRequest>

Accept a verification requests that we share with the given user with the given flow id.

This will move the verification request into the ready state.

Arguments

user_id - The ID of the user for which we would like to accept the verification requests.

flow_id - The ID that uniquely identifies the verification flow.

methods - A list of verification methods that we want to advertise as supported.

pub fn verification_request_content(
&self,
user_id: &str,
methods: Vec<String>
) -> Result<Option<String>, CryptoStoreError>

Get an m.key.verification.request content for the given user.

Arguments

user_id - The ID of the user which we would like to request to verify.

methods - The list of verification methods we want to advertise to support.

pub fn request_verification(
&self,
user_id: &str,
room_id: &str,
event_id: &str,
methods: Vec<String>
) -> Result<Option<VerificationRequest>, CryptoStoreError>

Request a verification flow to begin with the given user in the given room.

Arguments

user_id - The ID of the user which we would like to request to verify.

room_id - The ID of the room that represents a DM with the given user.

event_id - The event ID of the m.key.verification.request event that we sent out to request the verification to begin. The content for this request can be created using the verification_request_content() method.

methods - The list of verification methods we advertised as supported in the m.key.verification.request event.

pub fn request_verification_with_device(
&self,
user_id: &str,
device_id: &str,
methods: Vec<String>
) -> Result<Option<RequestVerificationResult>, CryptoStoreError>

Request a verification flow to begin with the given user’s device.

Arguments

user_id - The ID of the user which we would like to request to verify.

device_id - The ID of the device that we wish to verify.

methods - The list of verification methods we advertised as supported in the m.key.verification.request event.

pub fn request_self_verification(
&self,
methods: Vec<String>
) -> Result<Option<RequestVerificationResult>, CryptoStoreError>

Request a verification flow to begin with our other devices.

Arguments

methods - The list of verification methods we want to advertise to support.

pub fn get_verification(
&self,
user_id: &str,
flow_id: &str
) -> Option<Verification>

Get a verification flow object for the given user with the given flow id.

Arguments

user_id - The ID of the user for which we would like to fetch the verification.

flow_id - The ID that uniquely identifies the verification flow.

pub fn cancel_verification(
&self,
user_id: &str,
flow_id: &str,
cancel_code: &str
) -> Option<OutgoingVerificationRequest>

Cancel a verification for the given user with the given flow id using the given cancel code.

Arguments

user_id - The ID of the user for which we would like to cancel the verification.

flow_id - The ID that uniquely identifies the verification flow.

cancel_code - The error code for why the verification was cancelled, manual cancellatio usually happens with m.user cancel code. The full list of cancel codes can be found in the spec

pub fn confirm_verification(
&self,
user_id: &str,
flow_id: &str
) -> Result<Option<ConfirmVerificationResult>, CryptoStoreError>

Confirm a verification was successful.

This method should be called either if a short auth string should be confirmed as matching, or if we want to confirm that the other side has scanned our QR code.

Arguments

user_id - The ID of the user for which we would like to confirm the verification.

flow_id - The ID that uniquely identifies the verification flow.

pub fn start_qr_verification(
&self,
user_id: &str,
flow_id: &str
) -> Result<Option<QrCode>, CryptoStoreError>

Transition from a verification request into QR code verification.

This method should be called when one wants to display a QR code so the other side can scan it and move the QR code verification forward.

Arguments

user_id - The ID of the user for which we would like to start the QR code verification.

flow_id - The ID of the verification request that initiated the verification flow.

pub fn generate_qr_code(&self, user_id: &str, flow_id: &str) -> Option<String>

Generate data that should be encoded as a QR code.

This method should be called right before a QR code should be displayed, the returned data is base64 encoded (without padding) and needs to be decoded on the other side before it can be put through a QR code generator.

Note: You’ll need to call start_qr_verification() before calling this method, otherwise None will be returned.

Arguments

user_id - The ID of the user for which we would like to start the QR code verification.

flow_id - The ID that uniquely identifies the verification flow.

pub fn scan_qr_code(
&self,
user_id: &str,
flow_id: &str,
data: &str
) -> Option<ScanResult>

Pass data from a scanned QR code to an active verification request and transition into QR code verification.

This requires an active VerificationRequest to succeed, returns None if no VerificationRequest is found or if the QR code data is invalid.

Arguments

user_id - The ID of the user for which we would like to start the QR code verification.

flow_id - The ID of the verification request that initiated the verification flow.

data - The data that was extracted from the scanned QR code as an base64 encoded string, without padding.

pub fn start_sas_verification(
&self,
user_id: &str,
flow_id: &str
) -> Result<Option<StartSasResult>, CryptoStoreError>

Transition from a verification request into short auth string based verification.

Arguments

user_id - The ID of the user for which we would like to start the SAS verification.

flow_id - The ID of the verification request that initiated the verification flow.

pub fn start_sas_with_device(
&self,
user_id: &str,
device_id: &str
) -> Result<Option<StartSasResult>, CryptoStoreError>

Start short auth string verification with a device without going through a verification request first.

Note: This has been largely deprecated and the request_verification_with_device() method should be used instead.

Arguments

user_id - The ID of the user for which we would like to start the SAS verification.

device_id - The ID of device we would like to verify.

pub fn accept_sas_verification(
&self,
user_id: &str,
flow_id: &str
) -> Option<OutgoingVerificationRequest>

Accept that we’re going forward with the short auth string verification.

Arguments

user_id - The ID of the user for which we would like to accept the SAS verification.

flow_id - The ID that uniquely identifies the verification flow.

pub fn get_emoji_index(&self, user_id: &str, flow_id: &str) -> Option<Vec<i32>>

Get a list of emoji indices of the emoji representation of the short auth string.

Note: A SAS verification needs to be started and in the presentable state for this to return the list of emoji indices, otherwise returns None.

Arguments

user_id - The ID of the user for which we would like to get the short auth string.

flow_id - The ID that uniquely identifies the verification flow.

pub fn get_decimals(&self, user_id: &str, flow_id: &str) -> Option<Vec<i32>>

Get the decimal representation of the short auth string.

Note: A SAS verification needs to be started and in the presentable state for this to return the list of decimals, otherwise returns None.

Arguments

user_id - The ID of the user for which we would like to get the short auth string.

flow_id - The ID that uniquely identifies the verification flow.

pub fn bootstrap_cross_signing(
&self
) -> Result<BootstrapCrossSigningResult, CryptoStoreError>

Create a new private cross signing identity and create a request to upload the public part of it to the server.

pub fn cross_signing_status(&self) -> CrossSigningStatus

Get the status of the private cross signing keys.

This can be used to check which private cross signing keys we have stored locally.

pub fn export_cross_signing_keys(&self) -> Option<CrossSigningKeyExport>

Export all our private cross signing keys.

The export will contain the seed for the ed25519 keys as a base64 encoded string.

This method returns None if we don’t have any private cross signing keys.

pub fn import_cross_signing_keys(
&self,
export: CrossSigningKeyExport
) -> Result<(), SecretImportError>

Import our private cross signing keys.

The export needs to contain the seed for the ed25519 keys as a base64 encoded string.

pub fn enable_backup_v1(
&self,
key: MegolmV1BackupKey,
version: String
) -> Result<(), DecodeError>

Activate the given backup key to be used with the given backup version.

Warning: The caller needs to make sure that the given BackupKey is trusted, otherwise we might be encrypting room keys that a malicious party could decrypt.

The OlmMachine::verify_backup method can be used to so.

pub fn backup_enabled(&self) -> bool

Are we able to encrypt room keys.

This returns true if we have an active BackupKey and backup version registered with the state machine.

pub fn disable_backup(&self) -> Result<(), CryptoStoreError>

Disable and reset our backup state.

This will remove any pending backup request, remove the backup key and reset the backup state of each room key we have.

pub fn backup_room_keys(&self) -> Result<Option<Request>, CryptoStoreError>

Encrypt a batch of room keys and return a request that needs to be sent out to backup the room keys.

pub fn room_key_counts(&self) -> Result<RoomKeyCounts, CryptoStoreError>

Get the number of backed up room keys and the total number of room keys.

pub fn save_recovery_key(
&self,
key: Option<String>,
version: Option<String>
) -> Result<(), CryptoStoreError>

Store the recovery key in the cryptostore.

This is useful if the client wants to support gossiping of the backup key.

pub fn get_backup_keys(&self) -> Result<Option<BackupKeys>, CryptoStoreError>

Get the backup keys we have saved in our crypto store.

pub fn sign(&self, message: &str) -> HashMap<String, HashMap<String, String>>

Sign the given message using our device key and if available cross signing master key.

pub fn verify_backup(&self, auth_data: &str) -> Result<bool, CryptoStoreError>

Check if the given backup has been verified by us or by another of our devices that we trust.